overhaul resource tests
